+++ tests/integration/models/reddit/test_comment.py
@@ -1,6 +1,6 @@
 from asyncpraw.exceptions import ClientException, PRAWException
 from asyncpraw.models import Comment, Submission
-from prawcore import BadRequest
+from asyncprawcore import BadRequest
 import mock
 import pytest